Use array primitives also in BC interpreter
